requirements.txt
# Desktop Application Dependencies
tkinter==8.6
Pillow==8.2.0
openai==0.10.2
keyboard==0.13.5
pyperclip==1.8.2
python-dotenv==0.17.1
requests==2.25.1

# Web Application Dependencies
flask==2.1.0
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-bcrypt==0.7.1
gunicorn==20.1.0
email-validator==1.1.3
stripe==2.60.0

# Packaging
pyinstaller==4.5.1